% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ompBAM_imports.R
\name{idxstats}
\alias{idxstats}
\title{Runs idxstats}
\usage{
idxstats(bam_file, n_threads)
}
\arguments{
\item{bam_file}{The path to the bam file}

\item{n_threads}{The number of threads to use}
}
\value{
Output showing idxstats output
}
\description{
Runs idxstats
}
